Refactor constants list
