+++ app/javascript/utils/constant_files/wa_constants.js
